reignore vendor

+++ app/database/migrations/2013_12_10_172147_create_posts_table.php
@@ -15,7 +15,7 @@ class CreatePostsTable extends Migration {
 		Schema::create('posts', function(Blueprint $table)
 		{
 			$table->increments('id');
-			$table->integer('user_id')
+			$table->bigInteger('user_id')
 				->unsigned();
 			$table->text('goal');
 			$table->text('contact');
